Fix: UCL-INGI#5 - move create_a_publishing into Publishing class"
