+++ superform/superform/publishings.py
@@ -4,6 +4,28 @@ from superform.utils import login_required, datetime_converter, str_converter
 from superform.models import db, Publishing, Channel
 
 pub_page = Blueprint('publishings', __name__)
+
+
+def create_a_publishing(post, chn, form):
+    chan = str(chn.name)
+    title_post = form.get(chan + '_titlepost') if (form.get(chan + '_titlepost') is not None) else post.title
+    descr_post = form.get(chan + '_descriptionpost') if form.get(
+        chan + '_descriptionpost') is not None else post.description
+    link_post = form.get(chan + '_linkurlpost') if form.get(chan + '_linkurlpost') is not None else post.link_url
+    image_post = form.get(chan + '_imagepost') if form.get(chan + '_imagepost') is not None else post.image_url
+    date_from = datetime_converter(form.get(chan + '_datefrompost')) if datetime_converter(
+        form.get(chan + '_datefrompost')) is not None else post.date_from
+    date_until = datetime_converter(form.get(chan + '_dateuntilpost')) if datetime_converter(
+        form.get(chan + '_dateuntilpost')) is not None else post.date_until
+    pub = Publishing(post_id=post.id, channel_id=chn.id, state=0, title=title_post, description=descr_post,
+                     link_url=link_post, image_url=image_post,
+                     date_from=date_from, date_until=date_until)
+
+    db.session.add(pub)
+    db.session.commit()
+    return pub
+
+
 @pub_page.route('/moderate/<int:id>/<string:idc>',methods=["GET","POST"])
 @login_required()
 def moderate_publishing(id,idc):
